Day 10: Finish part 2
Took me a while to wrap my head around the angle math, wasn't nearly as
hard as I was making it out my mind. Really just needed to sleep on it.

Kind of happy with how the solution turned out. Definitely not optimized
for performance, but I think it's reasonably easy to follow the logic.