Day 18: Everything is fine
This might be some of the ugliest code I've ever written, but it gets
the job done! Took me ~4 hours and there was a point about halfway
through that I started to wish I had just built out the binary tree
structure instead of trying to manipulate the strings.

Oh well.