Day 19: Big oof!
Decided to wait on implementing this one for a few days because it
seemed exceptionally complex. Pretty happy with final solution because
it runs fairly quickly and the code isn't entirely unreadable.

Strugled for a while to implement the logic for finding the rotation
matrix using math until giving up on implementing SVD and just brute
force matching the possible rotational matrices we know are valid.