Day 24: Reading Assembly
My first pass on this was to try to brute force it by simply
implementing the ALU that parsed each instruction. That didn't finish
after running for ~6 hours so I took Hawth's advice and started reading
through the assembly by hand, which gave me `func` and a whole bunch of
stuff I could throw away from the input.

Afterwards, just some looping to generate all possible values of "z" as
we read through each chunk of instructions corresponding to a new digit
in the number. (It's too bad part 2 didn't involve the invalid "z"
values.) It's not fast, and runs in ~2m on my machine. But it gets
the right answers so it's good enough for now.